OTel: ensure cross-compilation support
While eBPF profiler is focused on Linux and amd64/arm64, the project is required
to be able to cross-compile on other OTel collector platforms as well.
