+++ .github/workflows/unit-test-on-pull-request.yml
@@ -62,6 +62,19 @@ jobs:
           make format
           git diff --exit-code || (echo 'Code base is not Go or C formatted, please run "make format" and commit the changes in this PR.'; exit 1)
 
+  otel-col-goos:
+    name: Crosscompile OTel Collector
+    runs-on: ubuntu-24.04
+    steps:
+      - name: Clone code
+        uses: actions/checkout@9c091bb21b7c1c1d1991bb908d89e4e9dddfe3e0 # v7.0.0
+      - name: Set up environment
+        uses: ./.github/workflows/env
+      - name: Cross compile
+        run: |
+          GOOS=windows go build ./collector/... 
+
+
   test:
     name: Test (${{ matrix.target_arch }}${{ matrix.go_flags }})
     runs-on: ${{ matrix.runner }}
