+++ collector/config/config.go
@@ -9,9 +9,7 @@ import (
 	"strings"
 	"time"
 
-	"go.opentelemetry.io/ebpf-profiler/internal/linux"
 	"go.opentelemetry.io/ebpf-profiler/interpreter/interpreterconfig"
-	"go.opentelemetry.io/ebpf-profiler/tracer"
 )
 
 const (
@@ -105,15 +103,6 @@ func (cfg *Config) Validate() error {
 		)
 	}
 
-	if cfg.ProbabilisticThreshold < 1 ||
-		cfg.ProbabilisticThreshold > tracer.ProbabilisticThresholdMax {
-		return fmt.Errorf(
-			"invalid argument for probabilistic-threshold. Value "+
-				"should be between 1 and %d",
-			tracer.ProbabilisticThresholdMax,
-		)
-	}
-
 	if cfg.OffCPUThreshold < 0.0 || cfg.OffCPUThreshold > 1.0 {
 		return errors.New(
 			"invalid argument for off-cpu-threshold. The value " +
@@ -126,19 +115,9 @@ func (cfg *Config) Validate() error {
 				"should be in the range [0..1]. 0 disables jitter")
 	}
 
-	if !cfg.NoKernelVersionCheck {
-		major, minor, patch, err := linux.GetCurrentKernelVersion()
-		if err != nil {
-			return fmt.Errorf("failed to get kernel version: %v", err)
-		}
-
-		var minMajor, minMinor uint32
-		minMajor, minMinor = 5, 10
-		if major < minMajor || (major == minMajor && minor < minMinor) {
-			return fmt.Errorf("host Agent requires kernel version "+
-				"%d.%d or newer but got %d.%d.%d", minMajor, minMinor, major, minor, patch)
-		}
+	if err := validateTracerConfig(cfg); err != nil {
+		return err
 	}
 
-	return nil
+	return validatePlatformConstraints(cfg)
 }
