+++ collector/controller_options.go
@@ -1,6 +1,8 @@
 // Copyright The OpenTelemetry Authors
 // SPDX-License-Identifier: Apache-2.0
 
+//go:build linux && (amd64 || arm64)
+
 package collector // import "go.opentelemetry.io/ebpf-profiler/collector"
 
 import (
